Mid-commit, working on flow-log
