+++ .github/agentic-flow/logs/v2.4.0.md
@@ -350,6 +350,21 @@ Blockers: scripts/coder-handoff-check.sh fails with "finalCheck not recorded as 
 - 22:10, Still rejected by TL even without Review as TL saw the plan was not cleaned up properly
 - 22:10 Stopped execution as failed - we need to address this plan hygiene issue before giving this feature 1 more try. I would try to move plan to floe-log as a part of the log, or can be a separate json. Also we need explicitely tell architect to revisit the whole approach after he fixes the issue to avoid cascading.
 
+## Run 4 (the same task)
+- Started at 23:40
+- 23:40 - passed to PM
+- 23:41, 7% context used - Story passed TL check, moved to Architect
+- Architect struggles to create new plan (but we verified that command exists, maybe it was not passed to Architect explicitely)
+- Architect looks like stuck - yes, ARchitect stuck how to initialize a plan (we are loosing time here)
+- 23:49 - Architect can't recover
+- 23:50, Architect manually created plan and populated it, passed to TL
+- TL also struggles with flow-log commands and reading and validating manually - clearly mismatch with usage of new tool.
+- TL directly returned plan to ARchitect without Review phase - smth new???
+- Plan created by Architect is complete trash - missing fields, no explanations on approach, maybe we should guide it how to fill in a plan or provide example plan to use for reference?
+- Architect again struggls with plan related commands - his commands are so huge that it's not maintainable, we need to give him opportunity to update in big batches, instead of field by filed. 
+- I think I'll stop execution here - it's a real flow blocker and we can't proceed until it fixed.
+- 23:56 - stopped execution as most time architect was struggling with a tool - tool needs to be tested in more relalistic flow and upgraded accordingly.
+
 
 ### Post Run Checks
